Abstract signal handling, graceful max execution time, idle timeouts and memory usage limits to base consumers - so that RPC consumers can use these.
