+++ Command/DynamicConsumerCommand.php
@@ -11,6 +11,7 @@
 namespace OldSound\RabbitMqBundle\Command;
 
 use Symfony\Component\Console\Input\InputArgument;
+use Symfony\Component\Console\Input\InputInterface;
 
 class DynamicConsumerCommand extends BaseConsumerCommand
 {
@@ -30,7 +31,7 @@ class DynamicConsumerCommand extends BaseConsumerCommand
         return 'old_sound_rabbit_mq.%s_dynamic';
     }
 
-    protected function initConsumer($input)
+    protected function initConsumer(InputInterface $input)
     {
         parent::initConsumer($input);
         $this->consumer->setContext($input->getArgument('context'));
